no_code_airflow_mvp_minio/requirements.txt
pydantic>=2.7
PyYAML>=6.0.1
Jinja2>=3.1.2

s3fs>=2024.6.1
